video-pipeline/requirements.txt
twelvelabs
google-genai
python-dotenv
